Undo big refactor of earlier
There were just too many issues. While a nice idea in theory, it ended
up being too much of a hassle.

Everything in bot.js can be hotloaded again, which makes certain aspects
of this whole thing easier.

Still going to be a few bugs. Making this many changes in one hit is not
a bug-inducing-free process.